requirements.txt
asgiref==3.4.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-simple-captcha==0.5.14
djangorestframework==3.12.4
idna==2.10
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.20
PyJWT==2.2.0
python3-openid==3.2.0
pytz==2021.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.4
